* main: (167 commits)
  fix: during compaction in single node mode, the uuids were different (#2705)
  fix: dynamic filter icon (#2700)
  fix: wrong notification on dashboard error (#2703)
  fix: added IAM roles test ids (#2696)
  fix: ofga large number of streams (#2695)
  chore(deps): bump thehanimo/pr-title-checker from 1.4.1 to 1.4.2 (#2694)
  fix: IAM UI fixes (#2692)
  fix: search around error when order by _timestamp asc (#2689)
  fix: fixed IAM Permissions issue (#2690)
  fix: router panic (#2688)
  fix: disable internal login (#2686)
  fix: search around error when order by _timestamp asc (#2687)
  fix: support native users in rbac (#2683)
  chore: rename ZO_MEM_FILE_MAX_SIZE to ZO_MAX_FILE_SIZE_IN_MEMORY (#2684)
  perf: optimize db connections (#2681)
  fix: remove xxhash (#2680)
  fix: 0.8.1-rc1 build issue (#2679)
  fix: need delete merged files from ingester cache db (#2677)
  feat: compactor supports lock on stream level (#2664)
  fix: dashboard geolocation issue (#2673)
  ...

+++ web/src/services/http.ts
@@ -14,19 +14,16 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import store from "../stores";
-import type { AxiosInstance } from 'axios';
+import type { AxiosInstance } from "axios";
 import axios from "axios";
 import config from "../aws-exports";
+import { Notify } from "quasar";
 
 const http = ({ headers } = {} as any) => {
-
   let instance: AxiosInstance;
   if (config.isEnterprise == "false") {
     headers = {
-      Authorization:
-        (config.isCloud == "true")
-          ? "Bearer " + localStorage.getItem("token")
-          : localStorage.getItem("token") || "",
+      Authorization: localStorage.getItem("access_token") || "",
       ...headers,
     };
     instance = axios.create({
@@ -36,17 +33,16 @@ const http = ({ headers } = {} as any) => {
     });
   } else {
     headers = {
-      Authorization: "Bearer " + localStorage.getItem("access_token"),
+      Authorization: localStorage.getItem("access_token"),
       ...headers,
     };
+
     instance = axios.create({
       // timeout: 10000,
       baseURL: store.state.API_ENDPOINT,
-      withCredentials: config.isEnterprise,
       headers,
     });
   }
-
 
   instance.interceptors.response.use(
     function (response) {
@@ -66,34 +62,73 @@ const http = ({ headers } = {} as any) => {
                 error.response.data["error"] || "Invalid credentials"
               )
             );
-            if ((config.isCloud == "true") && !error.request.responseURL.includes("/auth/login")) {
+            if (
+              config.isCloud == "true" &&
+              !error.request.responseURL.includes("/auth/login")
+            ) {
               store.dispatch("logout");
               localStorage.clear();
               sessionStorage.clear();
               window.location.reload();
             }
             // Check if the failing request is not the login or refresh token request
-            else if ((config.isEnterprise == "true") && !error.config.url.includes('/config/dex_login') && !error.config.url.includes('/config/dex_refresh')) {
+            else if (
+              config.isEnterprise == "true" &&
+              (store.state as any).zoConfig.dex_enabled &&
+              !error.config.url.includes("/config/dex_login") &&
+              !error.config.url.includes("/config/dex_refresh") &&
+              !error.config.url.includes("/auth/login")
+            ) {
               // Call refresh token API
-              const refreshToken = localStorage.getItem('refresh_token');
+              const refreshToken = localStorage.getItem("refresh_token");
 
               // Modify the request to include the refresh token
-              return instance.get('/config/dex_refresh', {
-                headers: { 'Authorization': `${refreshToken}` }
-              }).then(res => {
-                localStorage.setItem('access_token', res.data);
-                error.config.headers['Authorization'] = 'Bearer ' + res.data;
-                return instance(error.config);
-              }).catch(refreshError => {
-                store.dispatch('logout');
-                localStorage.clear();
-                sessionStorage.clear();
-                window.location.reload();
-                return Promise.reject(refreshError);
-              });
+              return instance
+                .get("/config/dex_refresh", {
+                  headers: { Authorization: `${refreshToken}` },
+                })
+                .then((res) => {
+                  localStorage.setItem("access_token", "Bearer " + res.data);
+                  error.config.headers["Authorization"] = "Bearer " + res.data;
+                  return instance(error.config);
+                })
+                .catch((refreshError) => {
+                  store.dispatch("logout");
+                  localStorage.clear();
+                  sessionStorage.clear();
+                  window.location.reload();
+                  return Promise.reject(refreshError);
+                });
             } else {
-              console.log(JSON.stringify(error.response.data["error"] || "Invalid credentials"));
+              console.log(
+                JSON.stringify(
+                  error.response.data["error"] || "Invalid credentials"
+                )
+              );
             }
+            break;
+          case 403:
+            if (config.isEnterprise == "true" || config.isCloud == "true") {
+              Notify.create({
+                message:
+                  "Unauthorized Access: You are not authorized to perform this operation, please contact your administrator.",
+                timeout: 0, // This ensures the notification does not close automatically
+                color: "negative", // Customize color as needed
+                position: "top",
+                actions: [
+                  {
+                    color: "white",
+                    icon: "close",
+                    size: "sm",
+                  },
+                ],
+              });
+            }
+            console.log(
+              JSON.stringify(
+                error.response.data["error"] || "Unauthorized Access"
+              )
+            );
             break;
           case 404:
             console.log(
